🚚 Rename: 한 폴더에 여러 종류의 컴포넌트가 존재할 경우, 각 컴포넌트의 폴더에 정리하도록 수정
