+++ src/pages/SharePage/SharePage.tsx
@@ -5,8 +5,8 @@ import * as S from "./SharePage.style";
 // Components
 import HeadNav from "src/Components/BasicLayout/HeadNav";
 import ShareFolderProfile from "src/pages/SharePage/Components/SharePageProfile";
-import LinkSearchBar from "src/Components/LinkCard/LinkSearchBar";
-import LinkCardCollection from "src/Components/LinkCard/LinkCardCollection";
+import LinkSearchBar from "src/Components/LinkCard/LInkSearchBar/LinkSearchBar";
+import LinkCardCollection from "src/Components/LinkCard/LinkCardCollection/LinkCardCollection";
 
 // Types
 import UserLinkDataType from "src/@types/UserLinkDataType";
